[MOVE/#214] 캘린더 엔티티 이동
