+++ data/src/main/java/com/terning/data/datasourceimpl/ScrapDataSourceImpl.kt
@@ -4,7 +4,7 @@ import com.terning.data.datasource.ScrapDataSource
 import com.terning.data.dto.NonDataBaseResponse
 import com.terning.data.dto.request.ScrapColorRequestDto
 import com.terning.data.service.ScrapService
-import com.terning.domain.entity.CalendarScrapRequest
+import com.terning.domain.entity.calendar.CalendarScrapRequest
 import javax.inject.Inject
 
 class ScrapDataSourceImpl @Inject constructor(
