+++ domain/src/main/java/com/terning/domain/repository/ScrapRepository.kt
@@ -1,6 +1,6 @@
 package com.terning.domain.repository
 
-import com.terning.domain.entity.CalendarScrapRequest
+import com.terning.domain.entity.calendar.CalendarScrapRequest
 
 interface ScrapRepository {
     suspend fun postScrap(calendarScrapRequest: CalendarScrapRequest): Result<Unit>
