+++ feature/src/main/java/com/terning/feature/calendar/calendar/component/CalendarDetailDialog.kt
@@ -45,7 +45,7 @@ import com.terning.core.designsystem.theme.TerningMain
 import com.terning.core.designsystem.theme.TerningTheme
 import com.terning.core.designsystem.theme.White
 import com.terning.core.extension.noRippleClickable
-import com.terning.domain.entity.CalendarScrapDetail
+import com.terning.domain.entity.calendar.CalendarScrapDetail
 import com.terning.feature.intern.component.InternInfoRow
 import java.time.LocalDate
