+++ feature/src/main/java/com/terning/feature/calendar/list/component/CalendarScrap.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.unit.dp
 import com.terning.core.designsystem.component.item.InternItem
 import com.terning.core.designsystem.component.item.ScrapBox
 import com.terning.core.extension.noRippleClickable
-import com.terning.domain.entity.CalendarScrapDetail
+import com.terning.domain.entity.calendar.CalendarScrapDetail
 
 @Composable
 fun CalendarScrap(
