+++ feature/src/main/java/com/terning/feature/calendar/month/component/CalendarMonth.kt
@@ -16,7 +16,7 @@ import com.terning.core.designsystem.theme.TerningPointTheme
 import com.terning.core.extension.getDateAsMapString
 import com.terning.core.extension.isToday
 import com.terning.core.extension.noRippleClickable
-import com.terning.domain.entity.CalendarScrap
+import com.terning.domain.entity.calendar.CalendarScrap
 import com.terning.feature.calendar.calendar.component.CalendarDay
 import com.terning.feature.calendar.month.model.MonthModel
 import java.time.LocalDate
