+++ feature/src/main/java/com/terning/feature/calendar/month/component/CalendarMonthScrap.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.terning.core.designsystem.theme.TerningTheme
 import com.terning.core.designsystem.theme.White
-import com.terning.domain.entity.CalendarScrap
+import com.terning.domain.entity.calendar.CalendarScrap
 import com.terning.feature.R
 
 private const val MAX_SCRAP_COUNT = 3
