+++ feature/src/main/java/com/terning/feature/calendar/month/model/CalendarMonthUiState.kt
@@ -1,7 +1,7 @@
 package com.terning.feature.calendar.month.model
 
 import com.terning.core.state.UiState
-import com.terning.domain.entity.CalendarScrap
+import com.terning.domain.entity.calendar.CalendarScrap
 
 data class CalendarMonthUiState(
     val loadState: UiState<Map<String, List<CalendarScrap>>> = UiState.Loading
