+++ feature/src/main/java/com/terning/feature/calendar/week/CalendarWeekViewModel.kt
@@ -14,8 +14,8 @@ import com.terning.core.designsystem.theme.CalPurple
 import com.terning.core.designsystem.theme.CalRed
 import com.terning.core.designsystem.theme.CalYellow
 import com.terning.core.state.UiState
-import com.terning.domain.entity.CalendarScrapDetail
-import com.terning.domain.entity.CalendarScrapRequest
+import com.terning.domain.entity.calendar.CalendarScrapDetail
+import com.terning.domain.entity.calendar.CalendarScrapRequest
 import com.terning.domain.repository.CalendarRepository
 import com.terning.domain.repository.ScrapRepository
 import com.terning.feature.R
