+++ feature/src/main/java/com/terning/feature/calendar/week/model/CalendarWeekUiState.kt
@@ -1,7 +1,7 @@
 package com.terning.feature.calendar.week.model
 
 import com.terning.core.state.UiState
-import com.terning.domain.entity.CalendarScrapDetail
+import com.terning.domain.entity.calendar.CalendarScrapDetail
 import java.time.LocalDate
 
 data class CalendarWeekUiState(
