+++ feature/src/main/java/com/terning/feature/search/searchprocess/SearchProcessViewModel.kt
@@ -3,7 +3,7 @@ package com.terning.feature.search.searchprocess
 import androidx.compose.ui.graphics.Color
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.terning.domain.entity.CalendarScrapRequest
+import com.terning.domain.entity.calendar.CalendarScrapRequest
 import com.terning.domain.entity.search.SearchResult
 import com.terning.domain.repository.ScrapRepository
 import com.terning.domain.repository.SearchRepository
